pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.115.14",
    "firecrawl-py>=2.15.0",
    "pydantic>=2.11.7",
    "pymongo>=4.13.2",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
